ui and darkmode fixes
