+++ components/ui/component-filters/deprecate-filter/deprecate-filter.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames';
 import React from 'react';
-import { Toggle } from '@teambit/design.ui.input.toggle';
+import { Toggle } from '@teambit/design.inputs.toggle-switch';
 import type { ComponentFilterCriteria } from '@teambit/component.ui.component-filters.component-filter-context';
 import { useComponentFilter } from '@teambit/component.ui.component-filters.component-filter-context';
 import styles from './deprecate-filter.module.scss';
