+++ workspace.jsonc
@@ -103,7 +103,7 @@
         "@teambit/bvm.path": "^0.1.2",
         "@teambit/capsule": "0.0.12",
         "@teambit/code.ui.code-compare-section": "0.0.7",
-        "@teambit/code.ui.code-tab-tree": "^0.0.619",
+        "@teambit/code.ui.code-tab-tree": "^0.0.620",
         "@teambit/code.ui.hooks.use-code-params": "0.0.497",
         "@teambit/code.ui.object-formatter": "0.0.1",
         "@teambit/code.ui.queries.get-component-code": "^0.0.503",
@@ -167,7 +167,6 @@
         "@teambit/design.inputs.input-text": "^1.1.4",
         "@teambit/design.inputs.selectors.checkbox-item": "1.2.0",
         "@teambit/design.inputs.selectors.menu-item": "^1.1.0",
-        "@teambit/design.inputs.selectors.multi-select": "0.0.20",
         "@teambit/design.inputs.text-area": "^0.0.20",
         "@teambit/design.inputs.toggle-button": "0.0.13",
         "@teambit/design.inputs.toggle-switch": "^0.0.8",
@@ -223,7 +222,7 @@
         "@teambit/evangelist.elements.icon": "1.0.5",
         "@teambit/evangelist.input.checkbox.label": "^1.1.0",
         "@teambit/evangelist.surfaces.dropdown": "1.0.3",
-        "@teambit/explorer.ui.command-bar": "^2.0.17",
+        "@teambit/explorer.ui.command-bar": "^2.0.18",
         "@teambit/explorer.ui.component-card": "^0.0.51",
         "@teambit/explorer.ui.gallery.component-grid": "0.0.496",
         "@teambit/git.modules.git-ignore": "^1.0.2",
